Workbench: labeled annotations
- Add support for creating and rendering labeled annotations
